Edits

+++ Fi.cpp
@@ -5,6 +5,6 @@ using namespace std;
 
 int main()
 {
-    const double Fi = (sqrt(5) + 1) / 2;
+    const float Fi = (sqrt(5) + 1) / 2;
     cout << Fi;
 }
